tests/miloto_tests/tests_pluggy.py
# Project : baloto-colombia
# File Name : tests_pluggy.py
# Dir Path : tests/miloto_tests
# Created on: 2025–06–14 at 01:34:17.

from __future__ import annotations

from typing import Sized
from typing import TYPE_CHECKING
from typing import TypeVar

import pytest
from pytest import param
from hamcrest import assert_that
from hamcrest import equal_to

if TYPE_CHECKING:
    pass

parameterize = pytest.mark.parametrize
SizedT = TypeVar("SizedT", bound=Sized)


def test_template() -> None:
    pass
